feat(trainer): :sparkles: enhance trainer commands with improved error handling and service annotations

+++ src/main/java/com/johan/gym_control/services/trainer/DeleteTrainerCommand.java
@@ -2,7 +2,9 @@ package com.johan.gym_control.services.trainer;
 
 import com.johan.gym_control.repositories.TrainerRepository;
 import com.johan.gym_control.services.interfaces.ICommandParametrized;
+import org.springframework.stereotype.Service;
 
+@Service
 public class DeleteTrainerCommand implements ICommandParametrized<Void, Long> {
   private final TrainerRepository trainerRepository;
 
@@ -12,6 +14,9 @@ public class DeleteTrainerCommand implements ICommandParametrized<Void, Long> {
 
   @Override
   public Void execute(Long id) {
+    if (!trainerRepository.existsById(id)) {
+      throw new IllegalArgumentException("Trainer with ID " + id + " does not exist.");
+    }
     trainerRepository.deleteById(id);
     return null;
   }
